finalized healing power

+++ Assets/Scripts/HealthBar.cs
@@ -5,5 +5,5 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetHealth(int health) => slider.value = health;
+    public void SetHealth(float health) => slider.value = health;
 }
